pyproject.toml
[tool.poetry]
name = "msal-bearer"
version = "1.3.3"
description = "Python package to get auth token interactively for a public client application using msal and msal-extension for caching."
authors = ["Åsmund Våge Fannemel <asmf@equinor.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/Equinor/msal-bearer"
classifiers = ["Operating System :: OS Independent"]

[tool.poetry.dependencies]
python = "^3.9.2"
msal = "^1.33.0"
msal-extensions = "^1.3.1"
cryptography = ">=43.0.3"
azure-identity = "^1.24.0"

[tool.poetry.group.dev.dependencies]
black = "^24.4.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
